+++ src/components/i18n/documentation/notes.md
