+++ src/components/i18n/i18n.stories.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text, boolean } from '@storybook/addon-knobs';
 import i18n from 'i18n-js';
-import notes from './notes.md';
+import notes from './documentation/notes.md';
+import Info from './documentation/Info';
 import I18nComponent from './i18n';
 import { StoryHeader, StoryCode } from '../../../.storybook/style/storybook-info.styles';
 
@@ -27,29 +28,7 @@ storiesOf('I18nComponent', module)
       />
     );
   }, {
-    info: {
-      text: (
-        <div>
-          <p>A widget for internationalisation of text.</p>
-
-          <StoryHeader>How to use an I18n component:</StoryHeader>
-
-          <p>In your file:</p>
-
-          <StoryCode padded>
-            {'import I18n from "carbon-react/lib/components/i18n";'}
-          </StoryCode>
-
-          <p>To render the message:</p>
-
-          <StoryCode padded>
-            {'<I18n scope="foo" />'}
-          </StoryCode>
-
-          <p>For additional properties specific to this component, see propTypes.</p>
-        </div>
-      )
-    },
+    info: { text: Info },
     notes: { markdown: notes },
     knobs: { escapeHTML: false }
   });
