+++ src/components/icon/documentation/notes.md
