+++ src/components/icon/icon.stories.js
@@ -2,7 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { text, select } from '@storybook/addon-knobs';
 import OptionsHelper from '../../utils/helpers/options-helper';
-import notes from './notes.md';
+import notes from './documentation/notes.md';
+import Info from './documentation/Info';
 import Icon from './icon';
 import { StoryHeader, StoryCode } from '../../../.storybook/style/storybook-info.styles';
 
@@ -40,34 +41,7 @@ storiesOf('Icon', module)
       />
     );
   }, {
-    info: {
-      text: (
-        <div>
-          <p>An Icon widget.</p>
-
-          <StoryHeader>How to use an Icon in a component:</StoryHeader>
-
-          <p>In your file</p>
-
-          <StoryCode padded>
-            {'import Icon from "carbon-react/lib/components/icon";'}
-          </StoryCode>
-
-          <p>To render an Icon:</p>
-
-          <StoryCode padded>
-            {'<Icon type="foo" />'}
-          </StoryCode>
-
-          <p>
-            {"'type' is a required prop"}
-          </p>
-
-          {/* eslint-disable-next-line */}
-          <p>This widget follows the <a href="https://facebook.github.io/react/blog/2015/09/10/react-v0.14-rc1.html#stateless-function-components">Stateless Functional Component</a> pattern.</p>
-        </div>
-      )
-    },
+    info: { text: Info },
     notes: { markdown: notes },
     knobs: { escapeHTML: false }
   });
