+++ src/components/menu/menu.stories.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { select } from '@storybook/addon-knobs';
 import OptionsHelper from '../../utils/helpers/options-helper';
-import notes from './notes.md';
+import notes from './documentation/notes.md';
 import { Menu, MenuItem, SubmenuBlock } from './menu';
 
 storiesOf('Menu', module)
